requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn
uvicorn
uvicorn todo_backend.asgi:application --host 0.0.0.0 --port 8000